refactor: move example handlers from lib.rs to examples/
- Create examples/ directory with separate example files
- Move GotoDefinitionHandler to examples/goto_definition.rs
- Move FileOpenHandler to examples/file_open.rs
- Move HoverHandler to examples/hover.rs
- Keep Location struct as public in lib.rs for reuse
- Remove ~110 lines of example code from lib.rs
- Each example now has its own main function and tests
- Maintain project's ~800 line constraint
